Migrate to minimal APIs
